Added the following features in GameEditor mode : List Items Add items Remove Items
The relation to the database is now properly done

Next to do :
Bug fix the Team add to database. Stat id bug
Code the acutal gameplay

+++ GameEditor/Methods/RemoveItems.cs
@@ -1,5 +1,4 @@
 ﻿using Castle.DynamicProxy;
-using Data;
 using Data.People;
 using Data.Store;
 using Data.Values;
@@ -19,7 +18,6 @@ namespace GameEditor.Methods
             { // Connect to database
                 Dictionary<int, Audience> audiences = new Dictionary<int, Audience>();
                 int cpt = 0;
-                int[] possibleAnswer = new int[context.audiences.Count() + 1];
                 int userInput = -1;
                 string message = "Which audience would you like to remove ?";
 
@@ -27,20 +25,27 @@ namespace GameEditor.Methods
                 { // Retrieve the audiences
                     audiences.Add(cpt, audience);
                     Console.WriteLine($"{cpt} - {audience.Name}");
-                    possibleAnswer[cpt] = cpt;
                     cpt++;
                 }
                 Console.WriteLine($"{context.audiences.Count()} - Go back to remove items menu");
-                possibleAnswer[context.audiences.Count()] = context.audiences.Count();
 
                 do
                 { // Ask which audience to remove
-                    userInput = Data.Methods.General.AskForUserInputInt(message, possibleAnswer);
-                } while (!audiences.ContainsKey(userInput));
-                Audience audienceToRemove = audiences[userInput];
+                    userInput = Data.Methods.General.AskForUserInputInt(message, 0, context.audiences.Count());
+                } while (!audiences.ContainsKey(userInput) && userInput != context.audiences.Count());
 
-                // Remove audience
-                context.audiences.Remove(audienceToRemove);
+                if (userInput != context.audiences.Count())
+                {
+                    Audience audienceToRemove = audiences[userInput];
+
+                    // Remove audience
+                    context.audiences.Remove(audienceToRemove);
+                    context.SaveChanges();
+                }
+                else
+                {
+                    General.RemoveItemsMenu();
+                }
             }
         }
 
@@ -50,7 +55,6 @@ namespace GameEditor.Methods
             { // Connect to database
                 Dictionary<int, Team> teams = new Dictionary<int, Team>();
                 int cpt = 0;
-                int[] possibleAnswer = new int[context.teams.Count() + 1];
                 int userInput = -1;
                 string message = "Which team would you like to remove ?";
 
@@ -58,20 +62,27 @@ namespace GameEditor.Methods
                 { // Retrieve the teams
                     teams.Add(cpt, team);
                     Console.WriteLine($"{cpt} - {team.Name}");
-                    possibleAnswer[cpt] = cpt;
                     cpt++;
                 }
                 Console.WriteLine($"{context.teams.Count()} - Go back to remove items menu");
-                possibleAnswer[context.teams.Count()] = context.teams.Count();
 
                 do
                 { // Ask which team to remove
-                    userInput = Data.Methods.General.AskForUserInputInt(message, possibleAnswer);
+                    userInput = Data.Methods.General.AskForUserInputInt(message, 0, context.teams.Count());
                 } while (!teams.ContainsKey(userInput));
-                Team teamToRemove = teams[userInput];
 
-                // Remove team
-                context.teams.Remove(teamToRemove);
+                if (userInput != context.audiences.Count())
+                {
+                    Team teamToRemove = teams[userInput];
+
+                    // Remove team
+                    context.teams.Remove(teamToRemove);
+                    context.SaveChanges();
+                }
+                else
+                {
+                    General.RemoveItemsMenu();
+                }
             }
         }
 
@@ -81,7 +92,6 @@ namespace GameEditor.Methods
             { // Connect to database
                 Dictionary<int, Player> players = new Dictionary<int, Player>();
                 int cpt = 0;
-                int[] possibleAnswer = new int[context.players.Count() + 1];
                 int userInput = -1;
                 string message = "Which player would you like to remove ?";
 
@@ -89,20 +99,27 @@ namespace GameEditor.Methods
                 { // Retrieve the players
                     players.Add(cpt, player);
                     Console.WriteLine($"{cpt} - {player.Name}");
-                    possibleAnswer[cpt] = cpt;
                     cpt++;
                 }
                 Console.WriteLine($"{context.players.Count()} - Go back to remove items menu");
-                possibleAnswer[context.players.Count()] = context.players.Count();
 
                 do
                 { // Ask which player to remove
-                    userInput = Data.Methods.General.AskForUserInputInt(message, possibleAnswer);
+                    userInput = Data.Methods.General.AskForUserInputInt(message, 0, context.players.Count());
                 } while (!players.ContainsKey(userInput));
-                Player playerToRemove = players[userInput];
 
-                // Remove team
-                context.players.Remove(playerToRemove);
+                if (userInput != context.audiences.Count())
+                {
+                    Player playerToRemove = players[userInput];
+
+                    // Remove team
+                    context.players.Remove(playerToRemove);
+                    context.SaveChanges();
+                }
+                    else
+                {
+                    General.RemoveItemsMenu();
+                }
             }
         }
 
@@ -112,7 +129,6 @@ namespace GameEditor.Methods
             { // Connect to database
                 Dictionary<int, Equipment> equipments = new Dictionary<int, Equipment>();
                 int cpt = 0;
-                int[] possibleAnswer = new int[context.equipments.Count() + 1];
                 int userInput = -1;
                 string message = "Which equipment would you like to remove ?";
 
@@ -120,20 +136,27 @@ namespace GameEditor.Methods
                 { // Retrieve the equipments
                     equipments.Add(cpt, equipment);
                     Console.WriteLine($"{cpt} - {equipment.Name}");
-                    possibleAnswer[cpt] = cpt;
                     cpt++;
                 }
                 Console.WriteLine($"{context.equipments.Count()} - Go back to remove items menu");
-                possibleAnswer[context.equipments.Count()] = context.equipments.Count();
 
                 do
                 { // Ask which equipment to remove
-                    userInput = Data.Methods.General.AskForUserInputInt(message, possibleAnswer);
+                    userInput = Data.Methods.General.AskForUserInputInt(message, 0, context.equipments.Count());
                 } while (!equipments.ContainsKey(userInput));
-                Equipment equipmentToRemove = equipments[userInput];
 
-                // Remove equipment
-                context.equipments.Remove(equipmentToRemove);
+                if (userInput != context.audiences.Count())
+                {
+                    Equipment equipmentToRemove = equipments[userInput];
+
+                    // Remove equipment
+                    context.equipments.Remove(equipmentToRemove);
+                    context.SaveChanges();
+                }
+                else
+                {
+                    General.RemoveItemsMenu();
+                }
             }
         }
 
@@ -143,7 +166,6 @@ namespace GameEditor.Methods
             { // Connect to database
                 Dictionary<int, Skill> skills = new Dictionary<int, Skill>();
                 int cpt = 0;
-                int[] possibleAnswer = new int[context.skills.Count() + 1];
                 int userInput = -1;
                 string message = "Which skill would you like to remove ?";
 
@@ -151,20 +173,27 @@ namespace GameEditor.Methods
                 { // Retrieve the skills
                     skills.Add(cpt, skill);
                     Console.WriteLine($"{cpt} - {skill.Name}");
-                    possibleAnswer[cpt] = cpt;
                     cpt++;
                 }
                 Console.WriteLine($"{context.skills.Count()} - Go back to remove items menu");
-                possibleAnswer[context.skills.Count()] = context.skills.Count();
 
                 do
                 { // Ask which skill to remove
-                    userInput = Data.Methods.General.AskForUserInputInt(message, possibleAnswer);
+                    userInput = Data.Methods.General.AskForUserInputInt(message, 0, context.skills.Count());
                 } while (!skills.ContainsKey(userInput));
-                Skill skillToRemove = skills[userInput];
 
-                // Remove skill
-                context.skills.Remove(skillToRemove);
+                if (userInput != context.audiences.Count())
+                {
+                    Skill skillToRemove = skills[userInput];
+
+                    // Remove skill
+                    context.skills.Remove(skillToRemove);
+                    context.SaveChanges();
+                }
+                else
+                {
+                    General.RemoveItemsMenu();
+                }
             }
         }
 
@@ -174,7 +203,6 @@ namespace GameEditor.Methods
             { // Connect to database
                 Dictionary<int, Stat> stats = new Dictionary<int, Stat>();
                 int cpt = 0;
-                int[] possibleAnswer = new int[context.stats.Count() + 1];
                 int userInput = -1;
                 string message = "Which stat would you like to remove ?";
 
@@ -182,20 +210,27 @@ namespace GameEditor.Methods
                 { // Retrieve the stats
                     stats.Add(cpt, stat);
                     Console.WriteLine($"{cpt} - {stat.Name}");
-                    possibleAnswer[cpt] = cpt;
                     cpt++;
                 }
                 Console.WriteLine($"{context.stats.Count()} - Go back to remove items menu");
-                possibleAnswer[context.stats.Count()] = context.stats.Count();
 
                 do
                 { // Ask which stat to remove
-                    userInput = Data.Methods.General.AskForUserInputInt(message, possibleAnswer);
+                    userInput = Data.Methods.General.AskForUserInputInt(message, 0, context.stats.Count());
                 } while (!stats.ContainsKey(userInput));
-                Stat statToRemove = stats[userInput];
 
-                // Remove stat
-                context.stats.Remove(statToRemove);
+                if (userInput != context.audiences.Count())
+                {
+                    Stat statToRemove = stats[userInput];
+
+                    // Remove stat
+                    context.stats.Remove(statToRemove);
+                    context.SaveChanges();
+                }
+                else
+                {
+                    General.RemoveItemsMenu();
+                }
             }
         }
 
@@ -205,7 +240,6 @@ namespace GameEditor.Methods
             { // Connect to database
                 Dictionary<int, Performance> performances = new Dictionary<int, Performance>();
                 int cpt = 0;
-                int[] possibleAnswer = new int[context.performances.Count() + 1];
                 int userInput = -1;
                 string message = "Which performance would you like to remove ?";
 
@@ -213,20 +247,27 @@ namespace GameEditor.Methods
                 { // Retrieve the performances
                     performances.Add(cpt, performance);
                     Console.WriteLine($"{cpt} - {performance.Name}");
-                    possibleAnswer[cpt] = cpt;
                     cpt++;
                 }
                 Console.WriteLine($"{context.performances.Count()} - Go back to remove items menu");
-                possibleAnswer[context.performances.Count()] = context.performances.Count();
 
                 do
                 { // Ask which performances to remove
-                    userInput = Data.Methods.General.AskForUserInputInt(message, possibleAnswer);
+                    userInput = Data.Methods.General.AskForUserInputInt(message, 0, context.performances.Count());
                 } while (!performances.ContainsKey(userInput));
-                Performance performanceToRemove = performances[userInput];
 
-                // Remove equipment
-                context.performances.Remove(performanceToRemove);
+                if (userInput != context.audiences.Count())
+                {
+                    Performance performanceToRemove = performances[userInput];
+
+                    // Remove equipment
+                    context.performances.Remove(performanceToRemove);
+                    context.SaveChanges();
+                }
+                else
+                {
+                    General.RemoveItemsMenu();
+                }
             }
         }
     }
